feat: change backoff strategy to exponential backoff
